+++ lib/WorseReflection/Tests/Integration/Core/Inference/NodeContextResolverTest.php
@@ -191,7 +191,7 @@ class NodeContextResolverTest extends IntegrationTestCase
                     }
 
                     EOT
-                , [], ['type' => 'string', 'value' => 'test']
+                , [], ['type' => 'string']
                 ];
 
         yield 'It returns the value of a method parameter with a constant' => [
@@ -208,7 +208,7 @@ class NodeContextResolverTest extends IntegrationTestCase
                     }
 
                     EOT
-                , [], ['type' => 'string', 'value' => 'test']
+                , [], ['type' => 'string']
                 ];
 
         yield 'It returns the FQN of a method parameter in an interface' => [
@@ -261,7 +261,7 @@ class NodeContextResolverTest extends IntegrationTestCase
                     }
 
                     EOT
-                , [], ['type' => 'string', 'value' => 'test']
+                , [], ['type' => 'string']
                 ];
 
         yield 'Ignores parameter on anonymous class' => [
@@ -531,7 +531,7 @@ class NodeContextResolverTest extends IntegrationTestCase
 
                     'bar<>';
                     EOT
-                , [], ['type' => '"bar"', 'value' => 'bar', 'symbol_type' => Symbol::STRING ]
+                , [], ['type' => '"bar"', 'symbol_type' => Symbol::STRING ]
                 ];
 
         yield 'It returns type for float' => [
@@ -540,7 +540,7 @@ class NodeContextResolverTest extends IntegrationTestCase
 
                     1.<>2;
                     EOT
-                , [], ['type' => '1.2', 'value' => 1.2, 'symbol_type' => Symbol::NUMBER],
+                , [], ['type' => '1.2', 'symbol_type' => Symbol::NUMBER],
                 ];
 
         yield 'It returns type for integer' => [
@@ -549,7 +549,7 @@ class NodeContextResolverTest extends IntegrationTestCase
 
                     12<>;
                     EOT
-                , [], ['type' => '12', 'value' => 12, 'symbol_type' => Symbol::NUMBER],
+                , [], ['type' => '12', 'symbol_type' => Symbol::NUMBER],
                 ];
 
         yield 'It returns type for octal integer' => [
@@ -558,7 +558,7 @@ class NodeContextResolverTest extends IntegrationTestCase
 
                     012<>;
                     EOT
-                , [], ['type' => '012', 'value' => 012, 'symbol_type' => Symbol::NUMBER],
+                , [], ['type' => '012', 'symbol_type' => Symbol::NUMBER],
                 ];
 
         yield 'It returns type for hexadecimal integer' => [
@@ -567,7 +567,7 @@ class NodeContextResolverTest extends IntegrationTestCase
 
                     0x1A<>;
                     EOT
-                , [], ['type' => '0x1A', 'value' => 0x1A, 'symbol_type' => Symbol::NUMBER],
+                , [], ['type' => '0x1A', 'symbol_type' => Symbol::NUMBER],
                 ];
 
         yield 'It returns type for binary integer' => [
@@ -576,7 +576,7 @@ class NodeContextResolverTest extends IntegrationTestCase
 
                     0b11<>;
                     EOT
-                , [], ['type' => '0b11', 'value' => 0b11, 'symbol_type' => Symbol::NUMBER],
+                , [], ['type' => '0b11', 'symbol_type' => Symbol::NUMBER],
                 ];
 
         yield 'It returns type for bool true' => [
@@ -585,7 +585,7 @@ class NodeContextResolverTest extends IntegrationTestCase
 
                     tr<>ue;
                     EOT
-                , [], ['type' => 'true', 'value' => true, 'symbol_type' => Symbol::BOOLEAN],
+                , [], ['type' => 'true', 'symbol_type' => Symbol::BOOLEAN],
                 ];
 
         yield 'It returns type for bool false' => [
@@ -594,7 +594,7 @@ class NodeContextResolverTest extends IntegrationTestCase
 
                     <>false;
                     EOT
-                , [], ['type' => 'false', 'value' => false, 'symbol_type' => Symbol::BOOLEAN],
+                , [], ['type' => 'false', 'symbol_type' => Symbol::BOOLEAN],
                 ];
 
         yield 'It returns type null' => [
@@ -603,8 +603,7 @@ class NodeContextResolverTest extends IntegrationTestCase
 
                     n<>ull;
                     EOT
-                , [], ['type' => 'null', 'value' => null],
-                ];
+                , [], ['type' => 'null',   ]             ];
 
         yield 'It returns type null case insensitive' => [
                 <<<'EOT'
@@ -612,8 +611,7 @@ class NodeContextResolverTest extends IntegrationTestCase
 
                     N<>ULL;
                     EOT
-                , [], ['type' => 'null', 'value' => null],
-                ];
+                , [], ['type' => 'null',   ]             ];
 
         yield 'It returns type and value for an array' => [
                 <<<'EOT'
@@ -621,7 +619,7 @@ class NodeContextResolverTest extends IntegrationTestCase
 
                     [ 'one' => 'two', 'three' => 3 <>];
                     EOT
-                , [], ['type' => 'array', 'value' => [ 'one' => 'two', 'three' => 3]],
+                , [], ['type' => 'array{one:"two",three:3}'],
                 ];
 
         yield 'Empty array' => [
@@ -630,8 +628,7 @@ class NodeContextResolverTest extends IntegrationTestCase
 
                     [  <>];
                     EOT
-                , [], ['type' => 'array', 'value' => [ ]],
-                ];
+                , [], ['type' => 'array{}']];
 
         yield 'It type for a class constant' => [
                 <<<'EOT'
@@ -774,7 +771,7 @@ class NodeContextResolverTest extends IntegrationTestCase
                     'barfoo' => NodeContext::for(
                         Symbol::fromTypeNameAndPosition(Symbol::STRING, 'barfoo', Position::fromStartAndEnd(0, 0))
                     )
-                    ->withType(TypeFactory::string())->withValue('hello')
+                    ->withType(TypeFactory::stringLiteral('hello'))
                 ], ['type' => 'string'],
             ];
 
@@ -911,7 +908,7 @@ class NodeContextResolverTest extends IntegrationTestCase
 
                     $barfoo ? <>'foobar' : 'barfoo';
                     EOT
-                , [], ['type' => '"foobar"', 'value' => 'foobar']
+                , [], ['type' => '"foobar"', ]
                 ];
 
         yield 'It uses condition value if ternery "if" is empty' => [
@@ -920,7 +917,7 @@ class NodeContextResolverTest extends IntegrationTestCase
 
                     'string' ?:<> new \stdClass();
                     EOT
-                , [], ['type' => '"string"', 'value' => 'string']
+                , [], ['type' => '"string"', ]
                 ];
 
         yield 'It returns unknown for ternary expressions with unknown condition values' => [
@@ -1194,9 +1191,6 @@ class NodeContextResolverTest extends IntegrationTestCase
                         $information->type()->__toString(),
                         $name,
                     );
-                    continue 2;
-                case 'value':
-                    $this->assertEquals($value, $information->value(), $name);
                     continue 2;
                 case 'symbol_type':
                     $this->assertEquals($value, $information->symbol()->symbolType(), $name);