+++ lib/WorseReflection/Tests/Integration/Core/SourceReflectorTest.php
@@ -59,7 +59,6 @@ class SourceReflectorTest extends IntegrationTestCase
 
         $offset = $this->createReflector($source)->reflectOffset($source, 27);
         $this->assertEquals('"Hello"', (string) $offset->symbolContext()->type());
-        $this->assertEquals('Hello', $offset->frame()->locals()->byName('$foobar')->first()->value());
     }
 
     /**